Large refactoring and bug fixes
The C code has been broken up into multiple files.  The following
changes have been made:

- Removal of ENTER/EXIT macros: to support this use case, the library
  should be compiled and run using a C-level profiler
- Compiler warnings fixed
- Fix a memory leak in the deallocation of curl_share objects
- Fix a null pointer bug in Eventloop_stop
- Introduce a helper function to allocate BufferNode objects
- Code read over, TODO comments added, etc.

On the python side:
- Move some info from setup.py to setup.cfg
- Update setup.py for the C code changes
