+++ setup.cfg
@@ -1,2 +1,24 @@
+[metadata]
+name = acurl
+description = An async Curl library.
+classifiers =
+    Development Status :: 3 - Alpha
+    Intended Audience :: Developers
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
+author = Tony Simpson
+author_email = agjasimpson@gmail.com
+maintainer = Sky Identity NFT team
+maintainer_email = aaron.ecay@sky.uk
+license = MIT
+url = https://github.com/sky-uk/acurl
+
+[options]
+packages = acurl
+setup_requires = pytest-runner
+install_requires = ujson
+tests_require = pytest
+
 [aliases]
 test=pytest
